queryable package clean-up save point
