+++ packages/nodejs/behaviors/proxy.ts
@@ -1,16 +1,16 @@
 import { TimelinePipe } from "@pnp/core";
-import { Queryable2 } from "@pnp/queryable";
+import { Queryable } from "@pnp/queryable";
 import { HttpsProxyAgent } from "https-proxy-agent";
 
-export function Proxy(proxyInit: string): TimelinePipe<Queryable2>;
+export function Proxy(proxyInit: string): TimelinePipe<Queryable>;
 // eslint-disable-next-line no-redeclare
-export function Proxy(proxyInit: any): TimelinePipe<Queryable2>;
+export function Proxy(proxyInit: any): TimelinePipe<Queryable>;
 // eslint-disable-next-line no-redeclare
-export function Proxy(proxyInit: any): TimelinePipe<Queryable2> {
+export function Proxy(proxyInit: any): TimelinePipe<Queryable> {
 
     const proxy = typeof proxyInit === "string" ? new HttpsProxyAgent(proxyInit) : proxyInit;
 
-    return (instance: Queryable2) => {
+    return (instance: Queryable) => {
 
         instance.on.pre(async (url, init, result) => {
