+++ packages/nodejs/behaviors/stream-parse.ts
@@ -1,7 +1,7 @@
 import { TimelinePipe } from "@pnp/core";
-import { parseBinderWithErrorCheck, Queryable2 } from "@pnp/queryable";
+import { parseBinderWithErrorCheck, Queryable } from "@pnp/queryable";
 
-export function StreamParse(): TimelinePipe<Queryable2> {
+export function StreamParse(): TimelinePipe<Queryable> {
 
     return parseBinderWithErrorCheck(async r => ({ body: r.body, knownLength: parseInt(r.headers.get("content-length"), 10) }));
 }
