+++ packages/queryable/behaviors/caching.ts
@@ -1,10 +1,10 @@
-import { Queryable2 } from "../queryable-2.js";
+import { Queryable } from "../queryable.js";
 import { isFunc, getHashCode, PnPClientStorage, dateAdd, TimelinePipe } from "@pnp/core";
 
 export type CacheKeyFactory = (url: string) => string;
 export type CacheExpireFunc = (url: string) => Date;
 
-export function Caching(store: "local" | "session" = "session", keyFactory?: CacheKeyFactory, expireFunc?: CacheExpireFunc): TimelinePipe<Queryable2> {
+export function Caching(store: "local" | "session" = "session", keyFactory?: CacheKeyFactory, expireFunc?: CacheExpireFunc): TimelinePipe<Queryable> {
 
     const storage = new PnPClientStorage();
     const s = store === "session" ? storage.session : storage.local;
@@ -18,10 +18,10 @@ export function Caching(store: "local" | "session" = "session", keyFactory?: Cac
         expireFunc = () => dateAdd(new Date(), "minute", 5);
     }
 
-    return (instance: Queryable2) => {
+    return (instance: Queryable) => {
         // Regardless of cached result, update cache async
         // instance.AsyncOverride = lazy;
-        instance.on.pre(async function (this: Queryable2, url: string, init: RequestInit, result: any): Promise<[string, RequestInit, any]> {
+        instance.on.pre(async function (this: Queryable, url: string, init: RequestInit, result: any): Promise<[string, RequestInit, any]> {
 
             const key = keyFactory(url.toString());
