+++ packages/queryable/behaviors/inject-headers.ts
@@ -1,9 +1,9 @@
 import { TimelinePipe } from "@pnp/core";
-import { Queryable2 } from "../queryable-2.js";
+import { Queryable } from "../queryable.js";
 
 export function InjectHeaders(headers: Record<string, string>): TimelinePipe {
 
-    return (instance: Queryable2) => {
+    return (instance: Queryable) => {
 
         instance.on.pre(async function (url: string, init: RequestInit, result: any) {
