+++ packages/queryable/behaviors/pnp-logging.ts
@@ -1,13 +1,13 @@
 import { TimelinePipe } from "@pnp/core";
 import { Logger, LogLevel } from "@pnp/logging";
-import { Queryable2 } from "../queryable-2.js";
+import { Queryable } from "../queryable.js";
 
-export function PnPLogging(activeLevel: LogLevel): TimelinePipe<Queryable2> {
+export function PnPLogging(activeLevel: LogLevel): TimelinePipe<Queryable> {
 
     // TODO: we set the active level here? or rework logger to be instance based for each behavior
     Logger.activeLogLevel = activeLevel;
 
-    return (instance: Queryable2) => {
+    return (instance: Queryable) => {
 
         instance.on.log(function (message: string, level: LogLevel) {
             Logger.write(message, level);
