+++ packages/queryable/index.ts
@@ -1,31 +1,17 @@
-export * from "./batch.js";
-export * from "./caching.js";
 export * from "./add-prop.js";
-export * from "./invokable-binder.js";
-export * from "./pipeline-binder.js";
+export * from "./invokable.js";
+export * from "./operations.js";
 export * from "./parsers.js";
-export * from "./pipeline.js";
 export * from "./queryable.js";
+export * from "./queryable-factory.js";
 export * from "./request-builders.js";
 
-export * from "./queryable-2.js";
-export * from "./operations.js";
-export * from "./parsers-2.js";
-export * from "./queryable-factory.js";
+/**
+ * Behavior exports
+ */
 export * from "./behaviors/browser-fetch.js";
 export * from "./behaviors/caching.js";
 export * from "./behaviors/cachingPessimistic.js";
 export * from "./behaviors/inject-headers.js";
 export * from "./behaviors/pnp-logging.js";
 export * from "./behaviors/queryable-from.js";
-export * from "./invokable.js";
-
-export {
-    extendGlobal,
-    extendObj,
-    extendFactory,
-    ExtensionType,
-    clearGlobalExtensions,
-    enableExtensions,
-    disableExtensions,
-} from "./invokable-extensions.js";
