+++ packages/queryable/queryable.ts
@@ -1,383 +1,235 @@
-import {
-    combine,
-    IFetchOptions,
-    IConfigOptions,
-    mergeOptions,
-    objectDefinedNotNull,
-    IRequestClient,
-    assign,
-    ILibraryConfiguration,
-    ITypedHash,
-    Runtime,
-    DefaultRuntime,
-    dateAdd,
-    stringIsNullOrEmpty,
-} from "@pnp/core";
-import { ICachingOptions } from "./caching.js";
-import { Batch } from "./batch.js";
-import { PipelineMethod } from "./pipeline.js";
-import { IODataParser, ODataParser } from "./parsers.js";
+import { combine, getGUID, Timeline, asyncReduce, broadcast, request, extendable } from "@pnp/core";
+import { LogLevel } from "@pnp/logging";
+import { IInvokable, invokable } from "./invokable.js";
 
-export function cloneQueryableData(source: Partial<IQueryableData>): Partial<IQueryableData> {
+export type QueryablePreObserver = (this: IQueryable2, url: string, init: RequestInit, result: any) => Promise<[string, RequestInit, any]>;
 
-    let body: any;
-    // this handles bodies that cannot be JSON encoded (Blob, etc)
-    // Note however, even bodies that can be serialized will not be cloned.
-    if (source.options && source.options.body) {
-        body = source.options.body;
-        source.options.body = "-";
-    }
+export type QueryableAuthObserver = (this: IQueryable2, url: URL, init: RequestInit) => Promise<[URL, RequestInit]>;
 
-    const s = JSON.stringify(source, (key: keyof IQueryableData, value: any) => {
+export type QueryableSendObserver = (this: IQueryable2, url: URL, init: RequestInit) => Promise<Response>;
 
-        switch (key) {
-            case "query":
-                return JSON.stringify([...(<Map<string, string>>value)]);
-            case "batch":
-            case "batchDependency":
-            case "cachingOptions":
-            case "clientFactory":
-            case "parser":
-                return "-";
-            default:
-                return value;
-        }
-    }, 0);
+export type QueryableParseObserver = (this: IQueryable2, url: URL, response: Response, result: any | undefined) => Promise<[URL, Response, any]>;
 
-    const parsed = JSON.parse(s, (key: keyof IQueryableData, value: any) => {
-        switch (key) {
-            case "query":
-                return new Map(JSON.parse(value));
-            case "batch":
-            case "batchDependency":
-            case "cachingOptions":
-            case "clientFactory":
-            case "parser":
-                return source[key];
-            default:
-                return value;
-        }
-    });
+export type QueryablePostObserver = (this: IQueryable2, url: URL, result: any | undefined) => Promise<[URL, any]>;
 
-    if (body) {
-        parsed.options.body = body;
-        source.options.body = body;
-    }
+export type QueryableDataObserver<T = any> = (this: IQueryable2, result: T) => void;
 
-    return parsed;
-}
+const DefaultMoments = {
+    pre: asyncReduce<QueryablePreObserver>(),
+    auth: asyncReduce<QueryableAuthObserver>(),
+    send: request<QueryableSendObserver>(),
+    parse: asyncReduce<QueryableParseObserver>(),
+    post: asyncReduce<QueryablePostObserver>(),
+    data: broadcast<QueryableDataObserver>(),
+} as const;
 
-export interface IQueryableData<DefaultActionType = any> {
-    batch: Batch | null;
-    batchIndex: number;
-    batchDependency: () => void | null;
-    cachingOptions: ICachingOptions | null;
-    cloneParentCacheOptions: ICachingOptions | null;
-    cloneParentWasCaching: boolean;
-    query: Map<string, string>;
-    options: IFetchOptions | null;
-    url: string;
-    parentUrl: string;
-    useCaching: boolean;
-    pipes?: PipelineMethod<DefaultActionType>[];
-    parser?: IODataParser<DefaultActionType>;
-    clientFactory?: () => IRequestClient;
-    method?: string;
-}
+@extendable()
+@invokable()
+export class Queryable<R> extends Timeline<typeof DefaultMoments> implements IQueryable2<R> {
 
-export interface OLD_IQueryable<DefaultActionType> {
-    data: Partial<IQueryableData<DefaultActionType>>;
-    query: Map<string, string>;
-    append(pathPart: string): void;
-    inBatch(batch: Batch): this;
-    addBatchDependency(): () => void;
-    toUrlAndQuery(): string;
-    toUrl(): string;
-    concat(pathPart: string): this;
-    configure(options: IConfigOptions): this;
-    configureFrom(o: OLD_IQueryable<DefaultActionType>): this;
-    usingCaching(options?: ICachingOptions): this;
-    usingParser(parser: IODataParser<any>): this;
-    withPipeline(pipeline: PipelineMethod<DefaultActionType>[]): this;
-    defaultAction(options?: IFetchOptions): Promise<DefaultActionType>;
-    getRuntime(): Runtime;
-    setRuntime(runtime: Runtime): this;
-    setRuntime(cloneGlobal: boolean, additionalConfig?: ITypedHash<any>): this;
-}
+    protected _url: string;
+    private _query: Map<string, string>;
 
-export abstract class OLD_Queryable<DefaultActionType = any> implements OLD_IQueryable<DefaultActionType> {
+    constructor(init: Queryable<any> | string, path?: string) {
 
-    private _data: Partial<IQueryableData<DefaultActionType>>;
-    private _runtime: Runtime;
+        let url = "";
+        let observers;
 
-    constructor(dataSeed: Partial<IQueryableData<DefaultActionType>> = {}) {
+        if (typeof init === "string") {
 
-        this._data = Object.assign({}, {
-            cloneParentWasCaching: false,
-            options: {},
-            parentUrl: "",
-            parser: new ODataParser<DefaultActionType>(),
-            query: new Map<string, string>(),
-            url: "",
-            useCaching: false,
-        }, cloneQueryableData(dataSeed));
-
-        this._runtime = null;
-    }
-
-    public get data(): Partial<IQueryableData<DefaultActionType>> {
-        return this._data;
-    }
-
-    public set data(value: Partial<IQueryableData<DefaultActionType>>) {
-        this._data = Object.assign({}, this.data, cloneQueryableData(value));
-    }
-
-    public getRuntime(): Runtime {
-
-        if (this._runtime === null) {
-            return DefaultRuntime;
-        }
-
-        return this._runtime;
-    }
-
-    public setRuntime(runtime: Runtime): this;
-    public setRuntime(cloneGlobal: boolean, additionalConfig?: ITypedHash<any>): this;
-    public setRuntime(...args: any[]): this {
-
-        // need to wait for ts update in spfx: [runtime: Runtime] | [cloneGlobal: boolean, additionalConfig?: ITypedHash<any>]
-
-        if (args[0] instanceof Runtime) {
-
-            this._runtime = args[0];
+            url = combine(init, path);
 
         } else {
 
-            this._runtime = args[0] ? new Runtime(DefaultRuntime.export()) : new Runtime();
+            const { _url } = init;
 
-            if (args.length > 1 && objectDefinedNotNull(args[1])) {
-                this._runtime.assign(args[1]);
-            }
+            url = combine(_url, path);
+            // TODO:: doesn't work due to data event (maybe others)
+            // pre, post, send, auth, error, log
+            // data
+            observers = init.observers;
         }
 
-        return this;
+        // TODO:: need to maybe filter out some handlers, like on data??
+        // need to trace through multiple objects inheriting and not
+        // TODO:: do we want state to be inherited?
+        super(DefaultMoments, observers, {});
+
+        this._url = url;
+        this._query = new Map<string, string>();
+        // process this one different
     }
 
     /**
-   * Gets the full url with query information
-   *
-   */
-    public abstract toUrlAndQuery(): string;
-
-    /**
-   * The default action for this
-   */
-    public abstract defaultAction(options?: IFetchOptions): Promise<DefaultActionType>;
-
-    /**
-  * Gets the current url
-  *
-  */
-    public toUrl(): string {
-        return this.data.url;
-    }
-
-    /**
-   * Directly concatenates the supplied string to the current url, not normalizing "/" chars
-   *
-   * @param pathPart The string to concatenate to the url
-   */
+    * Directly concatenates the supplied string to the current url, not normalizing "/" chars
+    *
+    * @param pathPart The string to concatenate to the url
+    */
     public concat(pathPart: string): this {
-        this.data.url += pathPart;
+        this._url += pathPart;
         return this;
     }
 
     /**
-   * Provides access to the query builder for this url
-   *
-   */
-    public get query(): Map<string, string> {
-        return this.data.query;
-    }
+     * Gets the full url with query information
+     *
+     */
+    public toRequestUrl(): string {
 
-    /**
-   * Sets custom options for current object and all derived objects accessible via chaining
-   *
-   * @param options custom options
-   */
-    public configure(options: IConfigOptions): this {
-        mergeOptions(this.data.options, options);
-        return this;
-    }
+        let u = this.toUrl();
 
-    /**
-   * Configures this instance from the configure options of the supplied instance
-   *
-   * @param o Instance from which options should be taken
-   */
-    public configureFrom(o: OLD_IQueryable<any>): this {
-
-        mergeOptions(this.data.options, o.data.options);
-
-        const sourceRuntime = o.getRuntime();
-        if (!sourceRuntime.get<{ "__isDefault__": boolean }, boolean>("__isDefault__")) {
-            this.setRuntime(sourceRuntime);
+        if (this._query.size > 0) {
+            u += "?" + Array.from(this._query).map((v: [string, string]) => `${v[0]}=${encodeURIComponent(v[1])}`).join("&");
         }
-        return this;
+
+        return u;
+    }
+
+    public get query(): Map<string, string> {
+        return this._query;
     }
 
     /**
-   * Enables caching for this request
-   *
-   * @param options Defines the options used when caching this request
-   */
-    public usingCaching(options?: string | ICachingOptions): this {
+     * Gets the current url
+     *
+     */
+    public toUrl(): string {
+        return this._url;
+    }
 
-        const runtime = this.getRuntime();
+    protected execute(requestInit: RequestInit = { method: "GET", headers: {} }): Promise<void> {
 
-        if (!runtime.get<ILibraryConfiguration, boolean>("globalCacheDisable")) {
+        setTimeout(async () => {
 
-            this.data.useCaching = true;
+            const requestId = getGUID();
+            let requestUrl: URL;
 
-            // handle getting just the key
-            if (typeof options === "string") {
-                if (stringIsNullOrEmpty(options)) {
-                    throw Error("Cache key cannot be empty.");
+            try {
+
+                this.log(`[id:${requestId}] Beginning request`, LogLevel.Info);
+
+                // eslint-disable-next-line prefer-const
+                let [url, init, result] = await this.emit.pre(this.toRequestUrl(), requestInit, undefined);
+
+                this.log(`[id:${requestId}] Url: ${url}`, LogLevel.Info);
+
+                if (typeof result !== "undefined") {
+
+                    this.log(`[id:${requestId}] Result returned from pre`, LogLevel.Info);
+                    this.log(`[id:${requestId}] Emitting data`, LogLevel.Verbose);
+                    this.emit.data(result);
+                    this.log(`[id:${requestId}] Emitted data`, LogLevel.Verbose);
+
+                    return;
                 }
-                options = <ICachingOptions>{ key: options };
+
+                this.log(`[id:${requestId}] Emitting auth`, LogLevel.Verbose);
+                [requestUrl, init] = await this.emit.auth(new URL(url), init);
+                this.log(`[id:${requestId}] Emitted auth`, LogLevel.Verbose);
+
+                this.log(`[id:${requestId}] Emitting send`, LogLevel.Verbose);
+                let response = await this.emit.send(requestUrl, init);
+                this.log(`[id:${requestId}] Emitted send`, LogLevel.Verbose);
+
+                this.log(`[id:${requestId}] Emitting parse`, LogLevel.Verbose);
+                [requestUrl, response, result] = await this.emit.parse(requestUrl, response, result);
+                this.log(`[id:${requestId}] Emitted parse`, LogLevel.Verbose);
+
+                this.log(`[id:${requestId}] Emitting post`, LogLevel.Verbose);
+                [requestUrl, result] = await this.emit.post(requestUrl, result);
+                this.log(`[id:${requestId}] Emitted post`, LogLevel.Verbose);
+
+                // TODO:: how do we handle the case where the request pipeline has worked as expected, however
+                // the result remains undefined? We shouldn't emit data as we don't have any, but should we have a
+                // completed event to signal the request is completed?
+                if (typeof result !== "undefined") {
+                    this.log(`[id:${requestId}] Emitting data`, LogLevel.Verbose);
+                    this.emit.data(result);
+                    this.log(`[id:${requestId}] Emitted data`, LogLevel.Verbose);
+                }
+
+            } catch (e) {
+
+                this.log(`[id:${requestId}] Emitting error: "${e.message || e}"`, LogLevel.Error);
+                // anything that throws we emit and continue
+                this.error(e);
+                this.log(`[id:${requestId}] Emitted error: "${e.message || e}"`, LogLevel.Error);
+
+            } finally {
+
+                this.log(`[id:${requestId}] Finished request`, LogLevel.Info);
             }
 
-            // this uses our local options if they are defined as defaults
-            const defaultOpts: Partial<ICachingOptions> = {
-                expiration: dateAdd(new Date(), "second", runtime.get<ILibraryConfiguration, number>("defaultCachingTimeoutSeconds")),
-                storeName: runtime.get<ILibraryConfiguration, "session" | "local">("defaultCachingStore"),
-            };
+        }, 0);
 
-            this.data.cachingOptions = assign(defaultOpts, options);
-        }
-
-        return this;
-    }
-
-    public usingParser(parser: IODataParser<any>): this {
-        this.data.parser = parser;
-        return this;
-    }
-
-    /**
-   * Allows you to set a request specific processing pipeline
-   *
-   * @param pipeline The set of methods, in order, to execute a given request
-   */
-    public withPipeline(pipeline: PipelineMethod<DefaultActionType>[]): this {
-        this.data.pipes = pipeline.slice(0);
-        return this;
-    }
-
-    /**
-   * Appends the given string and normalizes "/" chars
-   *
-   * @param pathPart The string to append
-   */
-    public append(pathPart: string): void {
-        this.data.url = combine(this.data.url, pathPart);
-    }
-
-    /**
-   * Adds this query to the supplied batch
-   *
-   * @example
-   * ```
-   *
-   * let b = pnp.sp.createBatch();
-   * pnp.sp.web.inBatch(b).get().then(...);
-   * b.execute().then(...)
-   * ```
-   */
-    public inBatch(batch: Batch): this {
-
-        if (this.hasBatch) {
-            throw Error("This query is already part of a batch.");
-        }
-
-        if (objectDefinedNotNull(batch)) {
-            batch.track(this);
-        }
-
-        return this;
-    }
-
-    /**
-   * Blocks a batch call from occuring, MUST be cleared by calling the returned function
-  */
-    public addBatchDependency(): () => void {
-        if (objectDefinedNotNull(this.data.batch)) {
-            return this.data.batch.addDependency();
-        }
-
-        return () => null;
-    }
-
-    /**
-   * Indicates if the current query has a batch associated
-   *
-   */
-    protected get hasBatch(): boolean {
-        return objectDefinedNotNull(this.data.batch);
-    }
-
-    /**
-   * The batch currently associated with this query or null
-   *
-   */
-    protected get batch(): Batch | null {
-        return this.hasBatch ? this.data.batch : null;
-    }
-
-    /**
-   * Gets the parent url used when creating this instance
-   *
-   */
-    protected get parentUrl(): string {
-        return this.data.parentUrl;
-    }
-
-    /**
-   * Clones this instance's data to target
-   *
-   * @param target Instance to which data is written
-   * @param settings [Optional] Settings controlling how clone is applied
-   */
-    protected cloneTo<T extends OLD_IQueryable<any>>(target: T, settings: { includeBatch?: boolean; includeQuery?: boolean } = {}): T {
-
-        // default values for settings
-        settings = assign({
-            includeBatch: true,
-            includeQuery: false,
-        }, settings);
-
-        target.data = Object.assign({}, cloneQueryableData(this.data), <Partial<IQueryableData<DefaultActionType>>>{
-            batch: null,
-            cloneParentCacheOptions: null,
-            cloneParentWasCaching: false,
-        }, cloneQueryableData(target.data));
-
-        target.configureFrom(this);
-
-        if (settings.includeBatch) {
-            target.inBatch(this.batch);
-        }
-
-        if (settings.includeQuery && this.query.size > 0) {
-            this.query.forEach((v, k) => target.query.set(k, v));
-        }
-
-        if (this.data.useCaching) {
-            target.data.cloneParentWasCaching = true;
-            target.data.cloneParentCacheOptions = this.data.cachingOptions;
-        }
-
-        return target;
+        return new Promise((resolve, reject) => {
+            this.on.data(resolve);
+            this.on.error(reject);
+        });
     }
 }
+
+// eslint-disable-next-line no-redeclare
+export interface Queryable<R = any> {
+    <T = R>(init?: RequestInit): Promise<T>;
+}
+
+// this interface is required to stop the class from recursively referencing itself through the DefaultBehaviors type
+export interface IQueryable2<R = any> extends Timeline<any>, IInvokable {
+    readonly query: Map<string, string>;
+    <T = R>(this: IQueryable2, init?: RequestInit): Promise<T>;
+    using(behavior: (intance: Timeline<any>) => Timeline<any>): this;
+    toRequestUrl(): string;
+    toUrl(): string;
+}
+
+/**
+* Directly concatenates the supplied string to the current url, not normalizing "/" chars
+*
+* @param pathPart The string to concatenate to the url
+*/
+// public concat(pathPart: string): this {
+//     this.data.url += pathPart;
+//     return this;
+// }
+
+
+
+/**
+* Clones this instance's data to target
+*
+* @param target Instance to which data is written
+* @param settings [Optional] Settings controlling how clone is applied
+*/
+// protected cloneTo<T extends IQueryable<any>>(target: T, settings: { includeBatch?: boolean; includeQuery?: boolean } = {}): T {
+
+//     // default values for settings
+//     settings = assign({
+//         includeBatch: true,
+//         includeQuery: false,
+//     }, settings);
+
+//     target.data = Object.assign({}, cloneQueryableData(this.data), <Partial<IQueryableData<DefaultActionType>>>{
+//         batch: null,
+//         cloneParentCacheOptions: null,
+//         cloneParentWasCaching: false,
+//     }, cloneQueryableData(target.data));
+
+//     target.configureFrom(this);
+
+//     if (settings.includeBatch) {
+//         target.inBatch(this.batch);
+//     }
+
+//     if (settings.includeQuery && this.query.size > 0) {
+//         this.query.forEach((v, k) => target.query.set(k, v));
+//     }
+
+//     if (this.data.useCaching) {
+//         target.data.cloneParentWasCaching = true;
+//         target.data.cloneParentCacheOptions = this.data.cachingOptions;
+//     }
+
+//     return target;
+// }
+
